third_party/blink/web_tests/wpt_internal/app-history/navigate-event/navigate-location.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<form id="form" action=""></form>
<script>
async_test(t => {
  appHistory.onnavigate = t.step_func_done(e => {
    assert_equals(e.canRespond, true);
    assert_equals(e.hashChange, true);
    assert_equals(e.formData, null);
  });
  location.href = "#";
}, "location API fires navigate event");
</script>
